# backend-ttgo/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Overvak - Sistema de monitoreo remoto para personas con Alzheimer a través de un wearable.",
    "description": "Documentación oficial de la API RESTful para el ecosistema **Overvak**. \n\nEste backend gestiona la comunicación entre los dispositivos wearables y la aplicación web. \n\n**Módulos del sistema:**\n1. **Autenticación:** Gestión de cuidadores y tokens.\n2. **Pacientes:** Vinculación de dispositivos y perfiles.\n3. **Datos IoT:** Ingesta de sensores, tales como el MAX30102, GPS (NEO6MV2) y acelerómetro.\n4. **Geocercas:** Algoritmos de punto en polígono para zonas seguras.\n5. **Alertas:** Notificaciones de emergencia.",
    "version": "1.2.1"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Introduce tu token así: Bearer <tu_token>"
    }
  },
  "paths": {
...
    "/api/datos/paciente/{pacienteId}": {
      "get": {
        "tags": [
          "Datos del paciente"
        ],
        "description": "Obtiene el historial completo de datos de un paciente específico",
        "parameters": [
          {
            "name": "pacienteId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del paciente (MongoDB ID)"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de datos obtenida"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Paciente no encontrado o no autorizado"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/pacientes/": {
      "post": {
        "tags": [
          "Pacientes"
        ],
        "description": "Registrar un nuevo paciente bajo el cuidado del usuario actual.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nombre": {
                  "type": "string",
                  "example": "Juan"
                },
                "apellidoP": {
                  "type": "string",
                  "example": "Pérez"
                },
                "apellidoM": {
                  "type": "string",
                  "example": "López"
                },
                "edad": {
                  "type": "number",
                  "example": 75
                },
                "dispositivo_id": {
                  "type": "string",
                  "example": "a1b2c3d4-1678886400"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Paciente registrado exitosamente"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Faltan campos obligatorios"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Conflicto: El ID del dispositivo ya está en uso por otro paciente."
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Pacientes"
        ],
        "description": "Obtiene la lista de todos los pacientes asociados al cuidador.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pacientes"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/pacientes/{id}/ubicacion": {
      "get": {
        "tags": [
          "Pacientes"
        ],
        "description": "Obtiene la última ubicación conocida (lat/long) de un paciente específico.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del paciente"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Ubicación encontrada",
            "schema": {
              "type": "object",
              "properties": {
                "latitud": {
                  "type": "number",
                  "example": 19.4326
                },
                "longitud": {
                  "type": "number",
                  "example": -99.1332
                },
                "fecha": {
                  "type": "string",
                  "example": "2025-05-26T10:00:00.000Z"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Paciente no encontrado o sin datos de ubicación"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/pacientes/{id}": {
      "put": {
        "tags": [
          "Pacientes"
        ],
        "description": "Actualiza la información del paciente. Nota: Para reasignar un dispositivo, primero libéralo del paciente anterior.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del paciente"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nombre": {
                  "type": "string",
                  "example": "Juanito"
                },
                "edad": {
                  "type": "number",
                  "example": 76
                },
                "dispositivo_id": {
                  "type": "string",
                  "example": "nuevo-id-o-null"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paciente actualizado"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflicto: El ID de dispositivo ya está en uso"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Pacientes"
        ],
        "description": "Elimina un paciente del sistema.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del paciente"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Eliminado correctamente"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
}